refactor: split into module
