+++ apps/storefront/app/lib/sanity/environment/index.ts
@@ -1,0 +1,2 @@
+export { createSanityEnvironment, useSanityEnvironment } from "./context";
+export { SanityEnvironment } from "./SanityEnvironment";
